api/requirements.txt
torch
torchvision
fastapi
uvicorn[standard]
pillow
python-multipart
numpy
